chore: #25 Moved index.ts files from component level to UI folder level
