--- src/src/features/ui/tile/index.ts
@@ -1,1 +1,0 @@
-﻿export * from "./tile";
